feat : end of the security config

See merge request s09-final/S09P31B205!17

Former-commit-id: b2e9f6afea8dbd0723bba88e6899eb8441c42937
Former-commit-id: e1d3a30b17cc5067024732cb2ac6c14b8a654165

+++ backend/src/main/java/com/ssafy/i5i/hotelAPI/domain/hotel/repository/AccommodationRepository.java
@@ -19,6 +19,6 @@ public interface AccommodationRepository extends JpaRepository<Accommodation, Lo
             "AS distance " +
             "FROM Accommodation " +  // FROM 절 추가
             "HAVING distance <= :distance", nativeQuery = true)
-    Optional<List<Accommodation>> findByCoordinate(@Param("latitude") BigDecimal latitude, @Param("longitude") BigDecimal longitude,
+    Optional<List<Accommodation>> findByCoordinate(@Param("latitude") Double latitude, @Param("longitude") Double longitude,
                                                   @Param("distance") Long distance);
 }
